refactor: move device constants into dedicated file, so constants.js does not rely on react-native and can be used from e2e-tests.
