+++ BrightID/src/components/Connections/SearchConnections.js
@@ -12,7 +12,7 @@ import { SvgXml } from 'react-native-svg';
 import { connect } from 'react-redux';
 import Ionicon from 'react-native-vector-icons/Ionicons';
 import Octicons from 'react-native-vector-icons/Octicons';
-import { DEVICE_LARGE, DEVICE_IOS } from '@/utils/constants';
+import { DEVICE_LARGE, DEVICE_IOS } from '@/utils/deviceConstants';
 import { navigate } from '@/NavigationService';
 import { setSearchParam } from '@/actions';
 import searchIcon from '@/static/search_icon.svg';
