+++ BrightID/src/components/Connections/TrustlevelSlider.js
@@ -4,7 +4,8 @@ import React from 'react';
 import { StyleSheet, View, Text } from 'react-native';
 
 import Slider from '@react-native-community/slider';
-import { connection_levels, WIDTH } from '@/utils/constants';
+import { connection_levels } from '@/utils/constants';
+import { WIDTH } from '@/utils/deviceConstants';
 import {
   connectionLevelColors,
   connectionLevelStrings,
