+++ BrightID/src/components/Helpers/EmptyList.js
@@ -3,7 +3,7 @@
 import React from 'react';
 import { Text, View, StyleSheet } from 'react-native';
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
-import { DEVICE_LARGE } from '@/utils/constants';
+import { DEVICE_LARGE } from '@/utils/deviceConstants';
 
 type Props = {
   title: string,
