+++ BrightID/src/components/Helpers/NotificationBanner.js
@@ -10,7 +10,9 @@ import groups from '@/static/add_group.svg';
 import connections from '@/static/add_person.svg';
 import misc from '@/static/trusted_connections.svg';
 import { setActiveNotification } from '@/actions';
-import { DEVICE_LARGE, CONNECTIONS_TYPE, HEIGHT } from '@/utils/constants';
+import { CONNECTIONS_TYPE } from '@/utils/constants';
+import { DEVICE_LARGE, HEIGHT } from '@/utils/deviceConstants';
+
 import { selectAllUnconfirmedConnections } from '@/components/PendingConnectionsScreens/pendingConnectionSlice';
 import DropDownAlertEnabled from '@/utils/DropDownAlertEnabler';
