+++ BrightID/src/components/OnboardingScreens/SignUp.js
@@ -17,7 +17,8 @@ import Spinner from 'react-native-spinkit';
 import { connect } from 'react-redux';
 import SimpleLineIcons from 'react-native-vector-icons/SimpleLineIcons';
 import { takePhoto, chooseImage } from '@/utils/images';
-import { DEVICE_LARGE, DEVICE_OS, ORANGE } from '@/utils/constants';
+import { ORANGE } from '@/utils/constants';
+import { DEVICE_LARGE, DEVICE_OS } from '@/utils/deviceConstants';
 import { handleBrightIdCreation } from './actions';
 import { checkTasks } from '../Tasks/TasksSlice';
