+++ BrightID/src/components/PendingConnectionsScreens/MyCodeScreen.js
@@ -17,7 +17,8 @@ import { SvgXml } from 'react-native-svg';
 import { useDispatch, useSelector } from 'react-redux';
 import Material from 'react-native-vector-icons/MaterialCommunityIcons';
 import ChannelSwitch from '@/components/Helpers/ChannelSwitch';
-import { DEVICE_LARGE, ORANGE } from '@/utils/constants';
+import { ORANGE } from '@/utils/constants';
+import { DEVICE_LARGE } from '@/utils/deviceConstants';
 import cameraIcon from '@/static/camera_icon_white.svg';
 import {
   channel_states,
