+++ BrightID/src/components/PendingConnectionsScreens/RatingButton.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import { StyleSheet, Text, TouchableOpacity } from 'react-native';
-import { DEVICE_LARGE } from '../../utils/constants';
+import { DEVICE_LARGE } from '@/utils/deviceConstants';
 
 type RatingButtonProps = {
   color: string,
