+++ BrightID/src/components/PendingConnectionsScreens/ReconnectView.js
@@ -4,8 +4,8 @@ import React, { useEffect, useState } from 'react';
 import { View, Text, StyleSheet, TouchableOpacity } from 'react-native';
 import moment from 'moment';
 import { useNavigation } from '@react-navigation/native';
+import { DEVICE_LARGE } from '@/utils/deviceConstants';
 import { ConnectionStats } from './ConnectionStats';
-import { DEVICE_LARGE } from '../../utils/constants';
 import { ProfileCard } from './ProfileCard';
 import {
   connectionLevelColors,
