+++ BrightID/src/components/Tasks/TaskCard.js
@@ -8,7 +8,7 @@ import {
   Linking,
 } from 'react-native';
 import React from 'react';
-import { DEVICE_LARGE } from '@/utils/constants';
+import { DEVICE_LARGE } from '@/utils/deviceConstants';
 import { TaskState } from './TaskState';
 
 type TaskCardProps = {
